+++ l3experimental/l3draw/build.lua
@@ -12,9 +12,5 @@ maindir = "../.."
 
 typesetfiles = {"l3draw.dtx", "l3draw-code.tex"}
 
--- Need color support
-checkdeps = {maindir .. "/l3backend", maindir .. "/l3kernel", maindir .. "/l3experimental/l3color"}
-typesetdeps = checkdeps
-
 -- Load the common build code
 dofile(maindir .. "/build-config.lua")